The original SCURK hooks spanning both the Primary and 1996 editions have been moved to a specific common source file, wrappers have also been added for various global vars and remote functions.
