+++ include/scurk_primary.h
@@ -48,36 +48,27 @@
 #endif
 
 // Program funcs
-SCURKPRIMARY_GAMECALL_MAIN(0x41083C, void, __cdecl, PlaceTileListDlg_DrawTile, TPlaceTileListDlg *, DRAWITEMSTRUCT *)
-SCURKPRIMARY_GAMECALL_MAIN(0x410A80, void, __cdecl, PlaceTileListDlg_ChangeHiLight, TPlaceTileListDlg *, DRAWITEMSTRUCT *)
-SCURKPRIMARY_GAMECALL_MAIN(0x4132EC, void, __cdecl, DebugOut, const char *, ...)
 SCURKPRIMARY_GAMECALL_MAIN(0x41378C, void *, __cdecl, gAllocBlock, size_t)
 SCURKPRIMARY_GAMECALL_MAIN(0x4137EC, void *, __cdecl, gResizeBlock, BYTE *, size_t)
 SCURKPRIMARY_GAMECALL_MAIN(0x413810, void, __cdecl, gFreeBlock, void *)
-SCURKPRIMARY_GAMECALL_MAIN(0x4139B4, int, __stdcall, gUpdateWaitWindow)
+SCURKPRIMARY_GAMECALL_MAIN(0x4139B4, void, __stdcall, gUpdateWaitWindow)
 SCURKPRIMARY_GAMECALL_MAIN(0x414D78, char *, __cdecl, EditableTileSet_GetLongName, cEditableTileSet *, int)
-SCURKPRIMARY_GAMECALL_MAIN(0x416F84, void, __cdecl, EditableTileSet_RenderShapeToTile, cEditableTileSet *, TBC45XDib *, int)
 SCURKPRIMARY_GAMECALL_MAIN(0x440F80, void, __cdecl, winscurkPlaceWindow_ClearCurrentTool, winscurkPlaceWindow *)
-SCURKPRIMARY_GAMECALL_MAIN(0x441CB8, void, __cdecl, winscurkPlaceWindow_HideTools, winscurkPlaceWindow *)
 SCURKPRIMARY_GAMECALL_MAIN(0x4495BC, void, __cdecl, winscurkMDIClient_RotateColors, winscurkMDIClient *, int)
 SCURKPRIMARY_GAMECALL_MAIN(0x457AF4, TBC45XPalette *, __cdecl, winscurkApp_GetPalette, winscurkApp *)
 SCURKPRIMARY_GAMECALL_MAIN(0x459604, void, __cdecl, winscurkApp_ScurkSound, winscurkApp *, int)
 SCURKPRIMARY_GAMECALL_MAIN(0x459CA0, winscurkPlaceWindow *, __cdecl, winscurkApp_GetPlaceWindow, winscurkApp *)
 
 // BC and WinAPI
-SCURKPRIMARY_GAMECALL_MAIN(0x45F691, TBC45XDC *, __cdecl, BCDC_Cons, TBC45XDC *, HDC)
-SCURKPRIMARY_GAMECALL_MAIN(0x45F718, void, __cdecl, BCDC_Dest, TBC45XDC *, char)
 SCURKPRIMARY_GAMECALL_MAIN(0x45F80F, void, __cdecl, BCDC_SelectObjectPalette, TBC45XDC *, TBC45XPalette *, int)
 SCURKPRIMARY_GAMECALL_MAIN(0x4695CA, unsigned int, __cdecl, BCWindow_DefaultProcessing, TBC45XParWindow *)
 SCURKPRIMARY_GAMECALL_MAIN(0x469C64, LRESULT, __cdecl, BCWindow_HandleMessage, TBC45XParWindow *, unsigned int msg, WPARAM wParam, LPARAM lParam)
-SCURKPRIMARY_GAMECALL_MAIN(0x46AE30, LRESULT, __cdecl, BCWindow_SendMessageA, TBC45XParWindow *, UINT msg, WPARAM wParam, LPARAM lParam)
 SCURKPRIMARY_GAMECALL_MAIN(0x46B108, void, __cdecl, BCWindow_SetCursor, TBC45XParWindow *, TBC45XModule *, const char *)
 SCURKPRIMARY_GAMECALL_MAIN(0x46B7C8, void, __cdecl, BCWindowDC_Destruct, TBC45XWindowDC *, char)
 SCURKPRIMARY_GAMECALL_MAIN(0x46B865, TBC45XClientDC *, __cdecl, BCClientDC_Construct, TBC45XClientDC *, HWND)
 SCURKPRIMARY_GAMECALL_MAIN(0x46F9DB, void, __cdecl, BCDialog_SetupWindow, TBC45XParDialog *)
 SCURKPRIMARY_GAMECALL_MAIN(0x46FABD, void, __cdecl, BCDialog_SetCaption, TBC45XParDialog *, char *)
 SCURKPRIMARY_GAMECALL_MAIN(0x46FB35, void, __cdecl, BCDialog_EvClose, TBC45XParDialog *)
-SCURKPRIMARY_GAMECALL_MAIN(0x4745B6, void, __cdecl, BCFrameWindow_EvSize, TBC45XParFrameWindow *, unsigned int, TBC45XSize *)
 SCURKPRIMARY_GAMECALL_MAIN(0x476475, int, __cdecl, BCListBox_GetSelIndex, TBC45XParListBox *)
 SCURKPRIMARY_GAMECALL_MAIN(0x477A62, LRESULT, __cdecl, BCListBox_GetString, TBC45XParListBox *, char *, int)
 SCURKPRIMARY_GAMECALL_MAIN(0x477A9B, int, __cdecl, BCListBox_SetItemData, TBC45XParListBox *, int, unsigned int)
@@ -88,8 +79,7 @@ SCURKPRIMARY_GAMECALL_MAIN(0x47854E, TBC45XMDIChild *, __cdecl, BCMDIClient_GetA
 SCURKPRIMARY_GAMEOFF_PTR(__int16,	wTileObjects,			0x491BAA)
 SCURKPRIMARY_GAMEOFF(WORD,			wColFastCnt,			0x498506)
 SCURKPRIMARY_GAMEOFF(WORD,			wColSlowCnt,			0x498508)
-SCURKPRIMARY_GAMEOFF(__int16,		wtoolNum,				0x4A482E)
-SCURKPRIMARY_GAMEOFF(__int16,	wtoolValue,					0x4A4866)
-SCURKPRIMARY_GAMEOFF(DWORD,		gReadyToGo,					0x4A6668)
+SCURKPRIMARY_GAMEOFF_PTR(__int16,	wtoolNum,				0x4A482E)
+SCURKPRIMARY_GAMEOFF(__int16,		wtoolValue,				0x4A4866)
 SCURKPRIMARY_GAMEOFF(winscurkApp *,	gScurkApplication,		0x4A6D88)
 SCURKPRIMARY_GAMEOFF_PTR(BC45Xstring,	TAppInitCmdLine,	0x4A73D0)
