+++ modules/scurkfix_1996.cpp
@@ -12,249 +12,31 @@
 
 static DWORD dwDummy;
 
-#define MISCHOOK_SCURK1996_DEBUG_INTERNAL 1
-#define MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE 2
-#define MISCHOOK_SCURK1996_DEBUG_PLACEANDCOPY 4
-
-#define MISCHOOK_SCURK1996_DEBUG DEBUG_FLAGS_NONE
-
-#ifdef DEBUGALL
-#undef MISCHOOK_SCURK1996_DEBUG
-#define MISCHOOK_SCURK1996_DEBUG DEBUG_FLAGS_EVERYTHING
-#endif
-
-UINT mischook_scurk1996_debug = MISCHOOK_SCURK1996_DEBUG;
-
 extern "C" void Hook_SCURK1996_winscurkMDIClient_CycleColors(winscurkMDIClient *pThis) {
-	TBC45XPalette *pPal;
-	TBC45XClientDC clDC;
-	TBC45XMDIChild *pMDIChild;
-	HWND hWndTargetOne, hWndTargetTwo, hWndTargetThree;
-	unsigned uFlags;
-	BOOL bRedraw, bNoChildren;
-
-	bRedraw = FALSE;
-	if (!IsIconic(pThis->pWnd->HWindow)) {
-		pPal = GameMain_winscurkApp_GetPalette_SCURK1996(gScurkApplication_SCURK1996);
-		GameMain_BCClientDC_Construct_SCURK1996(&clDC, pThis->pWnd->HWindow);
-		GameMain_BCDC_SelectObjectPalette_SCURK1996(&clDC, pPal, 0);
-		if (wColFastCnt_SCURK1996 == 5) {
-			GameMain_winscurkMDIClient_RotateColors_SCURK1996(pThis, 1);
-			AnimatePalette((HPALETTE)pPal->Handle, 0xAB, 0x31, pThis->mFastColors);
-			wColFastCnt_SCURK1996 = 0;
-			bRedraw = TRUE;
-		}
-		if (wColSlowCnt_SCURK1996 == 30) {
-			GameMain_winscurkMDIClient_RotateColors_SCURK1996(pThis, 0);
-			AnimatePalette((HPALETTE)pPal->Handle, 0xE0, 0x10, pThis->mSlowColors);
-			wColSlowCnt_SCURK1996 = 0;
-			bRedraw = TRUE;
-		}
-		++wColFastCnt_SCURK1996;
-		++wColSlowCnt_SCURK1996;
-		GameMain_BCWindowDC_Destruct_SCURK1996(&clDC, 0);
-
-		// Only call redraw if the given MDIChild is active, rather than
-		// refreshing all windows from pThis->pWnd->HWindow downwards.
-		//
-		// This reduces "a bit" of the flickering that was otherwise occurring
-		// across all windows; at this stage it is only limited to the active
-		// MDI Child.
-		if (bRedraw) {
-			pMDIChild = GameMain_BCMDIClient_GetActiveMDIChild_SCURK1996(pThis);
-			if (pMDIChild) {
-				bNoChildren = FALSE;
-				hWndTargetOne = 0;
-				hWndTargetTwo = 0;
-				hWndTargetThree = 0;
-				if (pMDIChild == (TBC45XMDIChild *)pThis->mPlaceWindow) {
-					hWndTargetOne = pThis->mPlaceWindow->__wndHead.pWnd->HWindow;
-					if (pThis->mPlaceWindow->pPlaceTileListDlg && pThis->mPlaceWindow->pPlaceTileListDlg->pListBox)
-						hWndTargetTwo = pThis->mPlaceWindow->pPlaceTileListDlg->pListBox->HWindow;
-				}
-				else if (pMDIChild == (TBC45XMDIChild *)pThis->mMoverWindow) {
-					if (pThis->mMoverWindow->pTileSourceWindow)
-						hWndTargetOne = pThis->mMoverWindow->pTileSourceWindow->HWindow;
-					if (pThis->mMoverWindow->pTileWorkingWindow)
-						hWndTargetTwo = pThis->mMoverWindow->pTileWorkingWindow->HWindow;
-				}
-				else if (pMDIChild == (TBC45XMDIChild *)pThis->mEditWindow) {
-					bNoChildren = TRUE;
-					hWndTargetOne = pThis->mEditWindow->__wndHead.pWnd->HWindow;
-					if (pThis->mEditWindow->pPaintWindow)
-						hWndTargetTwo = pThis->mEditWindow->pPaintWindow->HWindow;
-					if (pThis->mEditWindow->pPaletteWindow)
-						hWndTargetThree = pThis->mEditWindow->pPaletteWindow->HWindow;
-				}
-
-				uFlags = RDW_INVALIDATE;
-				if (!bNoChildren)
-					uFlags |= RDW_ALLCHILDREN;
-				if (hWndTargetOne)
-					RedrawWindow(hWndTargetOne, 0, 0, uFlags);
-				if (hWndTargetTwo)
-					RedrawWindow(hWndTargetTwo, 0, 0, RDW_ALLCHILDREN | RDW_INVALIDATE);
-				if (hWndTargetThree)
-					RedrawWindow(hWndTargetThree, 0, 0, RDW_ALLCHILDREN | RDW_INVALIDATE);
-			}
-		}
-	}
+	L_SCURK_winscurkMDIClient_CycleColors(pThis);
 }
 
 extern "C" void __cdecl Hook_SCURK1996_DebugOut(char const *fmt, ...) {
 	va_list args;
-	int len;
-	char* buf;
 
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_INTERNAL) == 0)
+	if ((mischook_scurk_debug & MISCHOOK_SCURK_DEBUG_INTERNAL) == 0)
 		return;
 
 	va_start(args, fmt);
-	len = _vscprintf(fmt, args) + 1;
-	buf = (char*)malloc(len);
-	if (buf) {
-		vsprintf_s(buf, len, fmt, args);
-
-		ConsoleLog(LOG_DEBUG, "0x%06X -> gDebugOut(): %s", _ReturnAddress(), buf);
-
-		free(buf);
-	}
-
+	L_SCURK_gDebugOut(fmt, args);
 	va_end(args);
 }
 
 extern "C" void __cdecl Hook_SCURK1996_PlaceTileListDlg_SetupWindow(TPlaceTileListDlg *pThis) {
-	char szTileStr[80 + 1];
-	int nItem, nMax;
-	int nIdx;
-	int iCXHScroll, imainRight, imainBottom, ilbCX, ilbCY;
-	TBC45XRect mainRect;
-
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "0x%06X -> PlaceTileListDlg_SetupWindow(0x%06X)\n", _ReturnAddress(), pThis);
-
-	strcpy_s(szTileStr, sizeof(szTileStr) - 1, "Tile");
-	GameMain_BCDialog_SetupWindow_SCURK1996(pThis);
-
-	iCXHScroll = GetSystemMetrics(SM_CXHSCROLL);
-
-	// First resize the dialogue.
-	GetClientRect(pThis->pWnd->HWindow, &mainRect);
-	imainRight = pThis->nMaxHitArea + iCXHScroll - mainRect.right;
-	imainBottom = pThis->nLBButtonWidth - mainRect.bottom;
-	GetWindowRect(pThis->pWnd->HWindow, &mainRect);
-	mainRect.right += imainRight + 8;
-	mainRect.bottom += imainBottom + 8;
-	SetWindowPos(pThis->pWnd->HWindow, HWND_TOP, mainRect.left, mainRect.top, mainRect.right - mainRect.left, mainRect.bottom - mainRect.top, SWP_NOZORDER | SWP_NOMOVE);
-
-	// Then resize the listbox control.
-	// If it is done in the wrong order it will fail "hard"
-	// on Windows 11 24H2+.
-	// Adjust the width and height slightly as well...
-	// otherwise it will still fail "hard".
-	ilbCX = (mainRect.right - mainRect.left) - 8;
-	ilbCY = (mainRect.bottom - mainRect.top) - 8;
-	GetClientRect(pThis->pWnd->HWindow, &mainRect);
-	SetWindowPos(pThis->pListBox->HWindow, HWND_TOP, mainRect.left + 3, mainRect.top + 3, ilbCX - 4, ilbCY + 2, SWP_NOZORDER);
-
-	GameMain_BCWindow_HandleMessage_SCURK1996(pThis->pListBox, LB_SETCOLUMNWIDTH, pThis->nMaxHitArea, 0);
-
-	nMax = wTileObjects_SCURK1996[3 * pThis->mNumTiles] + wTileObjects_SCURK1996[3 * pThis->mNumTiles + 1] - 1;
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "pThis->mNumTiles(%d), nMax(%d), pThis->nTileRow(%d)\n", pThis->mNumTiles, nMax, pThis->nTileRow);
-	for (nItem = wTileObjects_SCURK1996[3 * pThis->mNumTiles]; nMax > nItem; nItem += pThis->nTileRow) {
-		sprintf_s(szTileStr, sizeof(szTileStr) - 1, "Tile%04d%04d", nItem, nItem + pThis->nTileRow - 1);
-		nIdx = GameMain_BCListBox_AddString_SCURK1996(pThis->pListBox, szTileStr);
-		if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-			ConsoleLog(LOG_DEBUG, "nItem(%d), szTileStr[%s], nIdx(%d)\n", nItem, szTileStr, nIdx);
-		GameMain_BCListBox_SetItemData_SCURK1996(pThis->pListBox, nIdx, nItem);
-	}
+	L_SCURK_PlaceTileListDlg_SetupWindow(pThis);
 }
 
 extern "C" void __cdecl Hook_SCURK1996_PlaceTileListDlg_EvLButtonDblClk(TPlaceTileListDlg *pThis) {
-	int nCurSelRowIdx;
-	int nPosOne, nPosTwo;
-	char szBuf[80 + 1];
-	TBC45XPoint curPt;
-	TBC45XRect lbRect;
-
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "0x%06X -> PlaceTileListDlg_EvLButtonDblClk(0x%06X)\n", _ReturnAddress(), pThis);
-
-	nCurSelRowIdx = GameMain_BCListBox_GetSelIndex_SCURK1996(pThis->pListBox);
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "nCurSelRowIdx(%d)\n", nCurSelRowIdx);
-
-	GetCursorPos(&curPt);
-	GetWindowRect(pThis->pListBox->HWindow, &lbRect);
-	pThis->nXPos = (curPt.x - lbRect.left) / pThis->nPosWidth;
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "pThis->nXPos(%d)\n", pThis->nXPos);
-
-	GameMain_BCListBox_GetString_SCURK1996(pThis->pListBox, szBuf, nCurSelRowIdx);
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "szBuf(%s)\n", szBuf);
-
-	sscanf_s(szBuf, "Tile%04d%04d", &nPosOne, &nPosTwo);
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "nPosOne(%d), nPosTwo(%d)\n", nPosOne, nPosTwo);
-	pThis->nCurPos = pThis->nXPos + nPosOne;
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "pThis->nCurPos(%d)\n", pThis->nCurPos);
+	L_SCURK_PlaceTileListDlg_EvLButtonDblClk(pThis);
 }
 
 extern "C" void __cdecl Hook_SCURK1996_PlaceTileListDlg_EvLBNSelChange(TPlaceTileListDlg *pThis) {
-	int nCurSelRowIdx;
-	int nPosOne, nPosTwo;
-	int nValOne, nValTwo;
-	char szBuf[80 + 1];
-	char *pLongTileName;
-	winscurkPlaceWindow *pWindow;
-	TBC45XPoint curPt;
-	TBC45XRect lbRect;
-
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "0x%06X -> PlaceTileListDlg_EvLBNSelChange(0x%06X)\n", _ReturnAddress(), pThis);
-
-	nCurSelRowIdx = GameMain_BCListBox_GetSelIndex_SCURK1996(pThis->pListBox);
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "nCurSelRowIdx(%d)\n", nCurSelRowIdx);
-
-	GameMain_BCListBox_GetString_SCURK1996(pThis->pListBox, szBuf, nCurSelRowIdx);
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "szBuf(%s)\n", szBuf);
-
-	// These 3 lines have been added since in Windows 11 24H2-onwards
-	// it seems as if pThis[18] is not being set correctly.
-	// The following code is partially from the EvLButtonDblClk() call.
-	GetCursorPos(&curPt);
-	GetWindowRect(pThis->pListBox->HWindow, &lbRect);
-	pThis->nChldHndlorX = (curPt.x - lbRect.left);
-
-	nValOne = pThis->nChldHndlorX / pThis->nPosWidth;
-	sscanf_s(szBuf, "Tile%04d%04d", &nPosOne, &nPosTwo);
-	nValTwo = nValOne + nPosOne;
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PICKANDPLACE) != 0)
-		ConsoleLog(LOG_DEBUG, "nPosOne(%d), nPosTwo(%d), nValOne(%d), nValTwo(%d)\n", nPosOne, nPosTwo, nValOne, nValTwo);
-
-	if (nValTwo >= wTileObjects_SCURK1996[3 * pThis->mNumTiles + 1] + wTileObjects_SCURK1996[3 * pThis->mNumTiles]) {
-		GameMain_winscurkApp_ScurkSound_SCURK1996(gScurkApplication_SCURK1996, 3);
-		pThis->nSelected = 0;
-	}
-	else {
-		pThis->nXPos = nValOne;
-		pThis->nCurPos = nValTwo;
-		pThis->nSelected = 1;
-		pLongTileName = GameMain_EditableTileSet_GetLongName_SCURK1996(gScurkApplication_SCURK1996->mWorkingTiles, pThis->nCurPos);
-		GameMain_BCDialog_SetCaption_SCURK1996(pThis, pLongTileName);
-		wtoolValue_SCURK1996 = 8;
-		*(&wtoolNum_SCURK1996 + 8) = pThis->nCurPos;
-		InvalidateRect(pThis->pWnd->HWindow, 0, 0);
-		pWindow = GameMain_winscurkApp_GetPlaceWindow_SCURK1996(gScurkApplication_SCURK1996);
-		GameMain_winscurkPlaceWindow_ClearCurrentTool_SCURK1996(pWindow);
-		GameMain_BCWindow_SetCursor_SCURK1996(pWindow->__wndHead.pWnd, pThis->pWnd->Module, (const char *)30006);
-		GameMain_winscurkApp_ScurkSound_SCURK1996(gScurkApplication_SCURK1996, 1);
-	}
+	L_SCURK_PlaceTileListDlg_EvLBNSelChange(pThis);
 }
 
 extern "C" LONG __cdecl Hook_SCURK1996_EditableTileSet_mReadFromFile(cEditableTileSet *pThis, LPCSTR lpPathName) {
@@ -269,13 +51,7 @@ extern "C" void __declspec(naked) Hook_SCURK1996_MoverWindow_DisableMaximizeBox(
 		mov [pWnd], eax
 	}
 
-	if ((mischook_scurk1996_debug & MISCHOOK_SCURK1996_DEBUG_PLACEANDCOPY) != 0)
-		ConsoleLog(LOG_DEBUG, "0x%06X -> DisableMaximizeBox()\n", _ReturnAddress());
-
-	if (GetSystemMetrics(SM_CXSCREEN) > 700)
-		pWnd->Attr.Style &= ~WS_MAXIMIZEBOX;
-	else
-		pWnd->Attr.Style |= WS_MAXIMIZE;
+	pWnd = L_SCURK_MoverWindow_DisableMaximizeBox(pWnd);
 
 	__asm {
 		mov eax, pWnd
@@ -285,65 +61,7 @@ extern "C" void __declspec(naked) Hook_SCURK1996_MoverWindow_DisableMaximizeBox(
 }
 
 extern "C" void __cdecl Hook_SCURK1996_MoverWindow_EvGetMinMaxInfo(winscurkMoverWindow *pThis, MINMAXINFO *pMmi) {
-	LONG nCXScreen, x, y;
-
-	GameMain_BCWindow_DefaultProcessing_SCURK1996(pThis->__wndHead.pWnd);
-	nCXScreen = GetSystemMetrics(SM_CXSCREEN);
-	if (nCXScreen <= 640) {
-		x = 512;
-		y = 256;
-	}
-	else {
-		x = 640;
-		y = 480;
-	}
-
-	pMmi->ptMinTrackSize.x = x;
-	pMmi->ptMinTrackSize.y = y;
-
-	pMmi->ptMaxPosition.x = 0;
-	pMmi->ptMaxPosition.y = 0;
-	pMmi->ptMaxSize.x = x;
-	pMmi->ptMaxSize.y = y;
-
-	pMmi->ptMaxTrackSize.x = x;
-	pMmi->ptMaxTrackSize.y = y;
-}
-
-static char *L_SCURK1996_ProcessCmdLine(char *pMainPath, char *pCmdLineParms, BOOL *bValidFileEntry) {
-	int iArgc;
-	std::string str;
-	static char szFileArg[MAX_PATH + 1];
-
-	memset(szFileArg, 0, sizeof(szFileArg));
-
-	str = "\"";
-	str += pMainPath;
-	str += "\" ";
-	str += pCmdLineParms;
-
-	std::wstring wStr(str.begin(), str.end());
-	LPWSTR *pArgv;
-
-	pArgv = CommandLineToArgvW(wStr.c_str(), &iArgc);
-	if (pArgv) {
-		// When a drag-and-drop occurs (over the main program or a shortcut), the file argument
-		// is always at the very end; the processing will only accept that detail as well.
-		WideCharToMultiByte(CP_UTF8, 0, pArgv[iArgc - 1], -1, szFileArg, MAX_PATH, NULL, NULL);
-		_strlwr_s(szFileArg, sizeof(szFileArg) - 1);
-
-		free(pArgv);
-	}
-
-	*bValidFileEntry = FALSE;
-	if (iArgc > 1) {
-		if (strlen(szFileArg) > 0) {
-			if (L_IsPathValid(szFileArg))
-				*bValidFileEntry = TRUE;
-		}
-	}
-
-	return szFileArg;
+	L_SCURK_MoverWindow_EvGetMinMaxInfo(pThis, pMmi);
 }
 
 // And we're gritting our teeth...
@@ -357,16 +75,9 @@ extern "C" void __declspec(naked) __cdecl Hook_SCURK1996_OwlMainCommandLineFix(v
 		mov [pArgs], eax
 	}
 
-	BOOL bValidFileEntry;
 	char *pRet;
 
-	bValidFileEntry = FALSE;
-	if (nArgs >= 2)
-		pRet = L_SCURK1996_ProcessCmdLine(pArgs[0], TAppInitCmdLine_SCURK1996->p->array, &bValidFileEntry);
-	if (!bValidFileEntry)
-		pRet = NULL;
-
-	_chdir(szGamePath);
+	pRet = L_SCURK_OwlMainCommandLineFix(pArgs, nArgs);
 
 	__asm {
 		mov esi, [pRet]
@@ -375,20 +86,12 @@ extern "C" void __declspec(naked) __cdecl Hook_SCURK1996_OwlMainCommandLineFix(v
 }
 
 extern "C" void __cdecl Hook_SCURK1996_BCDialog_CmCancel(TBC45XDialog *pThis) {
-	winscurkPlaceWindow *pWindow;
-
-	// We really don't want to close the Place&Pick object selection
-	// dialogue by pressing escape...
-	pWindow = GameMain_winscurkApp_GetPlaceWindow_SCURK1996(gScurkApplication_SCURK1996);
-	if (pWindow && pWindow->pPlaceTileListDlg && pWindow->pPlaceTileListDlg == (TPlaceTileListDlg *)pThis)
-		return;
-
-	GameMain_BCDialog_EvClose_SCURK1996(pThis);
+	L_SCURK_BCDialog_CmCancel(pThis);
 }
 
 void InstallFixes_SCURK1996(void) {
 	if (mischook_debug == DEBUG_FLAGS_EVERYTHING)
-		mischook_scurk1996_debug = DEBUG_FLAGS_EVERYTHING;
+		mischook_scurk_debug = DEBUG_FLAGS_EVERYTHING;
 
 	InstallRegistryPathingHooks_SCURK1996();
 
@@ -417,6 +120,28 @@ void InstallFixes_SCURK1996(void) {
 	// This call is used to load the TILES.DB.
 	VirtualProtect((LPVOID)0x41510C, 5, PAGE_EXECUTE_READWRITE, &dwDummy);
 	NEWJMP((LPVOID)0x41510C, Hook_SCURK1996_EditableTileSet_mReadFromFile);
+
+	// 'nop' out the -1 case in the following functions in-regards to the
+	// maximum extent:
+	// - cPaintWindow::mZoomOut
+	// - cPaintWindow::mZoomIn
+	// - cPaintWindow::EvHScroll
+	VirtualProtect((LPVOID)0x443D28, 1, PAGE_EXECUTE_READWRITE, &dwDummy);
+	*(BYTE *)(0x443D28) = 0x90;
+	VirtualProtect((LPVOID)0x443D72, 1, PAGE_EXECUTE_READWRITE, &dwDummy);
+	*(BYTE *)(0x443D72) = 0x90;
+	VirtualProtect((LPVOID)0x443E58, 1, PAGE_EXECUTE_READWRITE, &dwDummy);
+	*(BYTE *)(0x443E58) = 0x90;
+	VirtualProtect((LPVOID)0x443E9D, 1, PAGE_EXECUTE_READWRITE, &dwDummy);
+	*(BYTE *)(0x443E9D) = 0x90;
+	VirtualProtect((LPVOID)0x446F76, 1, PAGE_EXECUTE_READWRITE, &dwDummy);
+	*(BYTE *)(0x446F76) = 0x90;
+
+	// Increased the maximum extent by 1 to fix the lack of the last
+	// right-side column of pixels:
+	// cPaintWindow::cPaintWindow
+	VirtualProtect((LPVOID)0x4432B4, 1, PAGE_EXECUTE_READWRITE, &dwDummy);
+	*(BYTE *)(0x4432B4) = 0x01;
 
 	// winscurkMoverWindow::EvSize():
 	// Temporarily remove the TFrameWindow::EvSize call.
