+++ sc2kfix.vcxproj
@@ -169,6 +169,7 @@ const char* szSC2KFixBuildInfo = "$(UserName)@$(ComputerName)  $([System.DateTim
     <ClCompile Include="modules\sc2k_1996.cpp" />
     <ClCompile Include="modules\sc2x.cpp" />
     <ClCompile Include="modules\scurkfix_1996.cpp" />
+    <ClCompile Include="modules\scurkfix_common.cpp" />
     <ClCompile Include="modules\scurkfix_primary.cpp" />
     <ClCompile Include="modules\settings.cpp" />
     <ClCompile Include="modules\spritebrowse_dialog.cpp" />
