+++ sc2kfix.vcxproj.filters
@@ -458,6 +458,9 @@
     <ClCompile Include="modules\lua_glue.cpp">
       <Filter>Source Files\Modules</Filter>
     </ClCompile>
+    <ClCompile Include="modules\scurkfix_common.cpp">
+      <Filter>Source Files\Modules</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="sc2kfix.rc">
